Separate WorkflowCombineTesting module
